Update README.md
In my code, what I did was I put my favorite movie and series, because both of them are the best I've ever watch so far. The only changes I did, is I changed the syntax that is double quoted into my favorite Movie as well as an another code for my favorite series. I also change the description, rating them from 1 to ten and beside it is my rating in number/int way. In addition I putted a term, for the description.